requirements.txt
Django==5.0.6
djangorestframework==3.15.0
djangorestframework-simplejwt==6.0.0
django-allauth==0.59.0
django-cors-headers==4.2.0
channels==4.1.0
channels-redis==4.0.0
daphne==4.1.2
django-redis==6.2.0
redis==5.3.0
python-dotenv==1.0.0
pytz==2025.7
